feat: eoa signMessage, typedData , connect
