+++ FRW/Modules/Browser/View/Auth/AuthnView.swift
@@ -298,120 +298,6 @@ extension AuthnView {
   }
 }
 
-struct AuthnAccountView: View {
-  let account: RNBridge.WalletAccount
-  let childAccounts: [RNBridge.WalletAccount]
-  var onTap: (() -> Void)?
-
-  init(
-    account: RNBridge.WalletAccount,
-    childAccounts: [RNBridge.WalletAccount] = [],
-    onTap: (() -> Void)? = nil
-  ) {
-    self.account = account
-    self.childAccounts = childAccounts
-    self.onTap = onTap
-  }
-
-  var body: some View {
-    Button(action: {
-      onTap?()
-    }) {
-      accountCardContent
-    }
-    .buttonStyle(PlainButtonStyle())
-  }
-  
-  private var accountCardContent: some View {
-    let content = HStack(spacing: 12) {
-      // Account avatar
-      WalletAvatarView(
-        emoji: .init(name: account.emojiInfo?.emoji),
-        avatar: account.avatar
-      )
-
-      // Account info
-      accountInfo
-
-      Spacer()
-
-      // Chevron
-      chevronIcon
-    }
-    
-    return content
-      .padding(.horizontal, 16)
-      .padding(.vertical, 10)
-      .frame(maxWidth: .infinity)
-      .background(Color.Brain.Light.lines5)
-      .cornerRadius(16)
-  }
-  
-  private var chevronIcon: some View {
-    Image(systemName: "chevron.right")
-      .font(.system(size: 14, weight: .medium))
-      .foregroundColor(.Brain.Core.icons)
-  }
-  
-  private var accountInfo: some View {
-    VStack(alignment: .leading, spacing: 2) {
-      // Name and address
-      nameAndAddressRow
-
-      // Balance
-      if let balance = account.balance {
-        Text("\(balance) FLOW")
-          .font(.inter(size: 12))
-          .foregroundColor(.Brain.Text.secondary)
-          .lineLimit(1)
-      }
-
-      // Child accounts indicators
-      if !childAccounts.isEmpty {
-        childAccountsRow
-      }
-    }
-  }
-  
-  private var nameAndAddressRow: some View {
-    HStack(spacing: 8) {
-      Text(account.name)
-        .font(.inter(size: 14, weight: .semibold))
-        .foregroundColor(.Brain.Text.primary)
-
-      Text("(\(account.address))")
-        .font(.inter(size: 12))
-        .foregroundColor(.Brain.Text.secondary)
-    }
-  }
-  
-  private var childAccountsRow: some View {
-    HStack(spacing: 4) {
-      // Parent indicator (if has children)
-      Image("account_link_mark")
-        .resizable()
-        .aspectRatio(contentMode: .fit)
-        .frame(width: 12, height: 12)
-        .padding(2)
-
-      // Child account avatars
-      childAccountAvatars
-    }
-  }
-  
-  private var childAccountAvatars: some View {
-    HStack(spacing: 4) {
-      ForEach(childAccounts, id: \.id) { child in
-        WalletAvatarView(
-          emoji: .init(name: child.emojiInfo?.name),
-          avatar: child.avatar,
-          size: .small
-        )
-      }
-    }
-  }
-}
-
 // MARK: - Helper Methods
 
 extension AuthnView {
@@ -422,48 +308,6 @@ extension AuthnView {
       return urlString
     }
     return host
-  }
-}
-
-// MARK: - Helper Extensions
-
-extension Color {
-  init(hex: String) {
-    let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-    var int: UInt64 = 0
-    Scanner(string: hex).scanHexInt64(&int)
-    let a, r, g, b: UInt64
-    switch hex.count {
-    case 3:  // RGB (12-bit)
-      (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-    case 6:  // RGB (24-bit)
-      (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-    case 8:  // ARGB (32-bit)
-      (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-    default:
-      (a, r, g, b) = (255, 0, 0, 0)
-    }
-    self.init(
-      .sRGB,
-      red: Double(r) / 255,
-      green: Double(g) / 255,
-      blue: Double(b) / 255,
-      opacity: Double(a) / 255
-    )
-  }
-}
-
-struct RoundedCorner: Shape {
-  var radius: CGFloat = .infinity
-  var corners: UIRectCorner = .allCorners
-
-  func path(in rect: CGRect) -> Path {
-    let path = UIBezierPath(
-      roundedRect: rect,
-      byRoundingCorners: corners,
-      cornerRadii: CGSize(width: radius, height: radius)
-    )
-    return Path(path.cgPath)
   }
 }
 
@@ -499,110 +343,4 @@ private struct TestAuthnData: AuthnDataProvider {
   }
 }
 
-#Preview("AuthnAccountView - Account Card") {
-  VStack(spacing: 20) {
-    // Main account with child accounts
-    AuthnAccountView(
-      account: RNBridge.WalletAccount(
-        id: "1",
-        name: "Panda",
-        address: "0x8888888888888ab",
-        emojiInfo: RNBridge.EmojiInfo(
-          emoji: "🐼",
-          name: "Panda",
-          color: "#D6D6D6"
-        ),
-        parentEmoji: nil,
-        parentAddress: nil,
-        avatar: nil,
-        isActive: true,
-        type: .main,
-        balance: "550.66",
-        nfts: nil
-      ),
-      childAccounts: [
-        RNBridge.WalletAccount(
-          id: "2",
-          name: "Penguin",
-          address: "0x123456",
-          emojiInfo: RNBridge.EmojiInfo(
-            emoji: "🐧",
-            name: "Penguin",
-            color: "#FFCB6C"
-          ),
-          parentEmoji: nil,
-          parentAddress: nil,
-          avatar: nil,
-          isActive: false,
-          type: .child,
-          balance: nil,
-          nfts: nil
-        ),
-        RNBridge.WalletAccount(
-          id: "3",
-          name: "Fox",
-          address: "0x789abc",
-          emojiInfo: RNBridge.EmojiInfo(
-            emoji: "🦊",
-            name: "Fox",
-            color: "#FFB6C1"
-          ),
-          parentEmoji: nil,
-          parentAddress: nil,
-          avatar: nil,
-          isActive: false,
-          type: .child,
-          balance: nil,
-          nfts: nil
-        ),
-      ],
-      onTap: {
-        print("Account tapped")
-      }
-    )
 
-    // Account without balance or children
-    AuthnAccountView(
-      account: RNBridge.WalletAccount(
-        id: "4",
-        name: "Lion",
-        address: "0xabcdef123456",
-        emojiInfo: RNBridge.EmojiInfo(
-          emoji: "🦁",
-          name: "Lion",
-          color: "#FFD700"
-        ),
-        parentEmoji: nil,
-        parentAddress: nil,
-        avatar: nil,
-        isActive: false,
-        type: .child,
-        balance: nil,
-        nfts: nil
-      )
-    )
-
-    // EVM account
-    AuthnAccountView(
-      account: RNBridge.WalletAccount(
-        id: "5",
-        name: "EVM Account",
-        address: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
-        emojiInfo: RNBridge.EmojiInfo(
-          emoji: "⚡",
-          name: "EVM",
-          color: "#627EEA"
-        ),
-        parentEmoji: nil,
-        parentAddress: nil,
-        avatar: nil,
-        isActive: false,
-        type: .evm,
-        balance: "1.23",
-        nfts: nil
-      )
-    )
-  }
-  .padding()
-  .background(Color.Brain.Core.background)
-}
